Resolved merge conflicts and refactored huffman-coding into multiple smaller files and functions.
